Move compression to internal module
